pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Aleх Bark <thrtth@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
gunicorn = "^22.0.0"
uvicorn = "^0.30.3"
django-bootstrap5 = "^24.2"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
